Backport move TestRng into casper_types.
