+++ node/Cargo.toml
@@ -102,6 +102,7 @@ vergen = "3"
 
 [dev-dependencies]
 assert-json-diff = "2.0.1"
+casper-types = { version = "1.5.0", path = "../types", features = ["datasize", "gens", "json-schema", "testing"] }
 fake_instant = "0.4.0"
 pnet = "0.28.0"
 pretty_assertions = "0.7.2"
