+++ node/src/components/event_stream_server/sse_server.rs
@@ -34,6 +34,8 @@ use warp::{
     Filter, Reply,
 };
 
+#[cfg(test)]
+use casper_types::testing::TestRng;
 use casper_types::{EraId, ExecutionEffect, ExecutionResult, ProtocolVersion, PublicKey};
 
 use super::DeployGetter;
@@ -41,7 +43,7 @@ use crate::types::{
     BlockHash, Deploy, DeployHash, FinalitySignature, JsonBlock, TimeDiff, Timestamp,
 };
 #[cfg(test)]
-use crate::{crypto::AsymmetricKeyExt, testing, testing::TestRng, types::Block};
+use crate::{crypto::AsymmetricKeyExt, testing, types::Block};
 
 /// The URL root path.
 pub const SSE_API_ROOT_PATH: &str = "events";
@@ -584,8 +586,10 @@ fn stream_to_client(
 mod tests {
     use std::iter;
 
+    use casper_types::testing::TestRng;
+
     use super::*;
-    use crate::{logging, testing::TestRng};
+    use crate::logging;
 
     async fn should_filter_out(
         event: &ServerSentEvent,
