+++ node/src/components/event_stream_server/tests.rs
@@ -10,6 +10,7 @@ use std::{
 };
 
 use assert_json_diff::assert_json_eq;
+use casper_types::testing::TestRng;
 use futures::{join, StreamExt};
 use http::StatusCode;
 use pretty_assertions::assert_eq;
@@ -25,7 +26,7 @@ use tokio::{
 use tracing::debug;
 
 use super::*;
-use crate::{logging, testing::TestRng};
+use crate::logging;
 use sse_server::{
     DeployAccepted, Id, QUERY_FIELD, SSE_API_DEPLOYS_PATH as DEPLOYS_PATH,
     SSE_API_MAIN_PATH as MAIN_PATH, SSE_API_ROOT_PATH as ROOT_PATH,
