+++ node/src/components/linear_chain/pending_signatures.rs
@@ -145,8 +145,8 @@ impl PendingSignatures {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{crypto::generate_ed25519_keypair, testing::TestRng, types::FinalitySignature};
-    use casper_types::EraId;
+    use crate::{crypto::generate_ed25519_keypair, types::FinalitySignature};
+    use casper_types::{testing::TestRng, EraId};
 
     use std::collections::BTreeMap;
