+++ node/src/components/storage/tests.rs
@@ -13,7 +13,9 @@ use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use smallvec::smallvec;
 
 use casper_hashing::Digest;
-use casper_types::{EraId, ExecutionResult, ProtocolVersion, PublicKey, SecretKey};
+use casper_types::{
+    testing::TestRng, EraId, ExecutionResult, ProtocolVersion, PublicKey, SecretKey,
+};
 
 use super::{
     construct_block_body_to_block_header_reverse_lookup, garbage_collect_block_body_v2_db,
@@ -27,7 +29,7 @@ use crate::{
         requests::{StateStoreRequest, StorageRequest},
         Multiple,
     },
-    testing::{ComponentHarness, TestRng, UnitTestEvent},
+    testing::{ComponentHarness, UnitTestEvent},
     types::{
         Block, BlockHash, BlockHeader, BlockSignatures, Deploy, DeployHash, DeployMetadata,
         DeployWithFinalizedApprovals, FinalitySignature, HashingAlgorithmVersion,
