+++ node/src/reactor/participating/tests.rs
@@ -12,6 +12,7 @@ use tokio::time;
 use casper_execution_engine::core::engine_state::GetBidsRequest;
 use casper_types::{
     system::auction::{Bids, DelegationRate},
+    testing::TestRng,
     EraId, Motes, ProtocolVersion, PublicKey, SecretKey, U512,
 };
 
@@ -31,9 +32,7 @@ use crate::{
         participating::{self, ParticipatingEvent},
         Reactor, ReactorExit, Runner,
     },
-    testing::{
-        self, filter_reactor::FilterReactor, network::Network, ConditionCheckReactor, TestRng,
-    },
+    testing::{self, filter_reactor::FilterReactor, network::Network, ConditionCheckReactor},
     types::{
         chainspec::{AccountConfig, AccountsConfig, ValidatorConfig},
         ActivationPoint, BlockHeader, Chainspec, Deploy, ExitCode, Timestamp,
