+++ node/src/types/chainspec/deploy_config.rs
@@ -15,9 +15,9 @@ use casper_types::{
     Motes, U512,
 };
 
-#[cfg(test)]
-use crate::testing::TestRng;
 use crate::types::TimeDiff;
+#[cfg(test)]
+use casper_types::testing::TestRng;
 
 #[derive(Copy, Clone, DataSize, PartialEq, Eq, Serialize, Deserialize, Debug)]
 // Disallow unknown fields to ensure config files and command-line overrides contain valid keys.
